feat: Refactor permissions and notifications, consolidate auth, and fix UI
- Add migrations for permissions split, assignments FK, and notification index pruning
- Refactor backend permissions and notification handlers
- Consolidate auth providers in shared and update frontends
- Fix Admin UI theme and Analytics page styles

+++ apps/backend/src/infrastructure/database/diesel_connection_manager.rs
@@ -149,7 +149,7 @@ impl DieselConnectionManager {
         Self::get_pool().await.ok().map(|pool| {
             let status = pool.status();
             DieselPoolStats {
-                size: status.size as usize,
+                size: status.size,
                 available: status.available as usize,
                 max_size: status.max_size,
             }
